Various architectural improvements:
- Separation of models from pages
- ChirpModel, which encapsulates helper functions
- Logs out when resetting tables to prevent errors
- Removed compiler warnings
