+++ src/Chirp.Web/Pages/Auth.cshtml
@@ -1,5 +1,5 @@
 @page "/login/oauth/auth"
-@model Chirp.Web.Pages.AuthModel
+@model Chirp.Web.Models.AuthModel
 
 @{
     Layout = null;
