+++ src/Chirp.Web/Pages/Public.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/"
-@model Chirp.Web.Pages.TimelineModel
+@model Chirp.Web.Models.TimelineModel
 @{
     ViewData["Title"] = "Chirp!";
     Layout = "Shared/_Layout";
