+++ src/Chirp.Web/Pages/ResetTables.cshtml
@@ -1,6 +1,6 @@
 @page "/reset"
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@model Chirp.Web.Pages.ResetTablesModel;
+@model Chirp.Web.Models.ResetTablesModel;
 
 @{
     Layout = "Shared/_Layout";
