+++ src/Chirp.Web/Pages/Shared/_PostCheep.cshtml
@@ -1,6 +1,6 @@
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 
-@if (User?.Identity != null && User.Identity?.IsAuthenticated == true && Model.IsUserOrPublicPage())
+@if (User?.Identity != null && User.Identity?.IsAuthenticated == true && Model.IsUserOrPublicPage(Model.RouteName))
 {
     <div class="cheepbox">
         <h3>What's on your mind, @(User?.Identity?.Name)?</h3>
