+++ src/Chirp.Web/Pages/UserTimeline.cshtml
@@ -1,6 +1,6 @@
 ﻿@page "/{author}"
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@model Chirp.Web.Pages.TimelineModel
+@model Chirp.Web.Models.TimelineModel
 
 @{
     ViewData["Title"] = "Chirp!";
